Revert refactoring of tests of SSH infrastructure (#3210)
