+++ rm/rm-server/src/test/java/functionaltests/RegressionTestSuite.java
@@ -28,8 +28,7 @@ package functionaltests;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
-import functionaltests.nodesource.TestLocalInfrastructureCronPolicy;
-import functionaltests.nodesource.TestSSHInfrastructureV2AllPolicies;
+import functionaltests.nodesource.TestSSHInfrastructureV2RestartDownNodesPolicy;
 import functionaltests.nodesrecovery.NodesRecoveryPropertyTest;
 import functionaltests.nodesrecovery.RecoverLocalInfrastructureTest;
 import functionaltests.nodesrecovery.RecoverSSHInfrastructureV2Test;
@@ -41,8 +40,7 @@ import functionaltests.selectionscript.SelectionScriptTimeOutTest;
 @RunWith(Suite.class)
 @Suite.SuiteClasses({ NonBlockingCoreTest.class, TestNSNodesPermissions.class, SelectionScriptTimeOutTest.class,
                       RecoverLocalInfrastructureTest.class, RecoverSSHInfrastructureV2Test.class,
-                      TestLocalInfrastructureCronPolicy.class, TestSSHInfrastructureV2AllPolicies.class,
-                      NodesRecoveryPropertyTest.class })
+                      TestSSHInfrastructureV2RestartDownNodesPolicy.class, NodesRecoveryPropertyTest.class })
 
 /**
  * @author ActiveEon Team
