+++ rm/rm-server/src/test/java/functionaltests/nodesource/TestLocalInfrastructureStaticPolicy.java
@@ -27,7 +27,6 @@ package functionaltests.nodesource;
 
 import static functionaltests.utils.RMTHelper.log;
 import static org.junit.Assert.*;
-import static org.ow2.proactive.utils.Lambda.repeater;
 
 import java.io.File;
 
@@ -42,8 +41,8 @@ import org.ow2.proactive.resourcemanager.nodesource.policy.StaticPolicy;
 import org.ow2.proactive.utils.FileToBytesConverter;
 import org.ow2.proactive.utils.NodeSet;
 
-import functionaltests.nodesource.helper.LocalInfrastructureTestHelper;
 import functionaltests.utils.RMFunctionalTest;
+import functionaltests.utils.RMTHelper;
 
 
 /**
@@ -61,7 +60,7 @@ public class TestLocalInfrastructureStaticPolicy extends RMFunctionalTest {
         byte[] creds = FileToBytesConverter.convertFileToByteArray(new File(PAResourceManagerProperties.getAbsolutePath(PAResourceManagerProperties.RM_CREDS.getValueAsString())));
         rmHelper.getResourceManager().createNodeSource(sourceName,
                                                        LocalInfrastructure.class.getName(),
-                                                       LocalInfrastructureTestHelper.getParameters(0),
+                                                       new Object[] { creds, 0, RMTHelper.DEFAULT_NODES_TIMEOUT, "" },
                                                        StaticPolicy.class.getName(),
                                                        null,
                                                        NODES_NOT_RECOVERABLE);
@@ -73,12 +72,13 @@ public class TestLocalInfrastructureStaticPolicy extends RMFunctionalTest {
         // creating node source
         // first parameter of im is empty default rmHelper url
         byte[] creds = FileToBytesConverter.convertFileToByteArray(new File(PAResourceManagerProperties.getAbsolutePath(PAResourceManagerProperties.RM_CREDS.getValueAsString())));
-        rmHelper.getResourceManager().createNodeSource(sourceName,
-                                                       LocalInfrastructure.class.getName(),
-                                                       LocalInfrastructureTestHelper.getParameters(defaultDescriptorNodesNb),
-                                                       StaticPolicy.class.getName(),
-                                                       null,
-                                                       NODES_NOT_RECOVERABLE);
+        rmHelper.getResourceManager()
+                .createNodeSource(sourceName,
+                                  LocalInfrastructure.class.getName(),
+                                  new Object[] { creds, defaultDescriptorNodesNb, RMTHelper.DEFAULT_NODES_TIMEOUT, "" },
+                                  StaticPolicy.class.getName(),
+                                  null,
+                                  NODES_NOT_RECOVERABLE);
 
         rmHelper.waitForNodeSourceCreation(sourceName, defaultDescriptorNodesNb, rmHelper.getMonitorsHandler());
     }
@@ -146,13 +146,17 @@ public class TestLocalInfrastructureStaticPolicy extends RMFunctionalTest {
 
         resourceManager.removeNodeSource(source1, true);
         //wait the n events of the n nodes removals of the node source
-        repeater.accept(defaultDescriptorNodesNb, () -> this.rmHelper.waitForAnyNodeEvent(RMEventType.NODE_REMOVED));
+        for (int i = 0; i < defaultDescriptorNodesNb; i++) {
+            rmHelper.waitForAnyNodeEvent(RMEventType.NODE_REMOVED);
+        }
 
         //wait for the event of the node source removal
         rmHelper.waitForNodeSourceEvent(RMEventType.NODESOURCE_REMOVED, source1);
 
         resourceManager.removeNodeSource(source2, true);
-        repeater.accept(defaultDescriptorNodesNb, () -> this.rmHelper.waitForAnyNodeEvent(RMEventType.NODE_REMOVED));
+        for (int i = 0; i < defaultDescriptorNodesNb; i++) {
+            rmHelper.waitForAnyNodeEvent(RMEventType.NODE_REMOVED);
+        }
 
         //wait for the event of the node source removal
         rmHelper.waitForNodeSourceEvent(RMEventType.NODESOURCE_REMOVED, source2);
