+++ rm/rm-server/src/test/java/functionaltests/nodesource/TestNodeSourceAfterRestart.java
@@ -23,7 +23,7 @@
  * If needed, contact us to obtain a release under GPL Version 2 or 3
  * or a different license than the AGPL.
  */
-package functionaltests.nodesource.deployment;
+package functionaltests.nodesource;
 
 import static functionaltests.utils.RMTHelper.log;
