+++ rm/rm-server/src/test/java/functionaltests/nodesrecovery/RecoverSSHInfrastructureV2Test.java
@@ -45,8 +45,7 @@ import org.ow2.proactive.resourcemanager.nodesource.infrastructure.SSHInfrastruc
 import org.ow2.proactive.resourcemanager.nodesource.policy.StaticPolicy;
 
 import functionaltests.monitor.RMMonitorEventReceiver;
-import functionaltests.nodesource.helper.SSHInfrastructureV2TestHelper;
-import functionaltests.nodesource.helper.StaticPolicyTestHelper;
+import functionaltests.nodesource.TestSSHInfrastructureV2;
 import functionaltests.utils.RMFunctionalTest;
 import functionaltests.utils.RMTHelper;
 
@@ -56,8 +55,6 @@ import functionaltests.utils.RMTHelper;
  * @since 22/06/17
  */
 public class RecoverSSHInfrastructureV2Test extends RMFunctionalTest {
-
-    private static final int NB_NODES = 3;
 
     private static final String START_CONFIG = "/functionaltests/config/functionalTRMProperties-RM-start-clean-db-nodes-recovery-enabled.ini";
 
@@ -69,7 +66,7 @@ public class RecoverSSHInfrastructureV2Test extends RMFunctionalTest {
 
     @Before
     public void setup() throws Exception {
-        SSHInfrastructureV2TestHelper.startSSHServer();
+        TestSSHInfrastructureV2.startSSHServer();
         RMTHelper.log("SSH server started");
         startRmAndCheckInitialState();
     }
@@ -78,7 +75,7 @@ public class RecoverSSHInfrastructureV2Test extends RMFunctionalTest {
     public void tearDown() throws Exception {
         // kill the remaining nodes that were preserved for the test
         RecoverInfrastructureTestHelper.killNodesWithStrongSigKill();
-        SSHInfrastructureV2TestHelper.stopSSHServer();
+        TestSSHInfrastructureV2.stopSSHServer();
     }
 
     @Test
@@ -114,17 +111,19 @@ public class RecoverSSHInfrastructureV2Test extends RMFunctionalTest {
 
         resourceManager.createNodeSource(NODE_SOURCE_NAME,
                                          SSHInfrastructureV2.class.getName(),
-                                         SSHInfrastructureV2TestHelper.getParameters(NB_NODES),
+                                         TestSSHInfrastructureV2.infraParams,
                                          StaticPolicy.class.getName(),
-                                         StaticPolicyTestHelper.getParameters(),
+                                         TestSSHInfrastructureV2.policyParameters,
                                          NODES_RECOVERABLE);
-        RMTHelper.waitForNodeSourceCreation(NODE_SOURCE_NAME, NB_NODES, this.rmHelper.getMonitorsHandler());
+        RMTHelper.waitForNodeSourceCreation(NODE_SOURCE_NAME,
+                                            TestSSHInfrastructureV2.NB_NODES,
+                                            this.rmHelper.getMonitorsHandler());
 
         RMMonitorEventReceiver resourceManagerMonitor = (RMMonitorEventReceiver) resourceManager;
         List<RMNodeSourceEvent> nodeSourceEvent = resourceManagerMonitor.getInitialState().getNodeSourceEvents();
         assertThat(nodeSourceEvent.size()).isEqualTo(1);
         assertThat(nodeSourceEvent.get(0).getSourceName()).isEqualTo(NODE_SOURCE_NAME);
-        assertThat(resourceManagerMonitor.getState().getAllNodes().size()).isEqualTo(NB_NODES);
+        assertThat(resourceManagerMonitor.getState().getAllNodes().size()).isEqualTo(TestSSHInfrastructureV2.NB_NODES);
     }
 
     private void restartRmAndCheckFinalState(boolean nodesShouldBeRecreated) throws Exception {
@@ -144,16 +143,16 @@ public class RecoverSSHInfrastructureV2Test extends RMFunctionalTest {
 
         // wait for nodes to be recreated if needed
         if (nodesShouldBeRecreated) {
-            rmHelper.waitForAnyMultipleNodeEvent(RMEventType.NODE_STATE_CHANGED, NB_NODES);
+            rmHelper.waitForAnyMultipleNodeEvent(RMEventType.NODE_STATE_CHANGED, TestSSHInfrastructureV2.NB_NODES);
         }
 
         // the nodes should have been recovered too, and should be alive
         Set<String> allNodes = resourceManagerMonitor.getState().getAllNodes();
-        assertThat(allNodes.size()).isEqualTo(NB_NODES);
+        assertThat(allNodes.size()).isEqualTo(TestSSHInfrastructureV2.NB_NODES);
         Set<String> nodeSourceNames = new HashSet<>();
         nodeSourceNames.add(NODE_SOURCE_NAME);
         Set<String> aliveNodeUrls = resourceManager.listAliveNodeUrls(nodeSourceNames);
-        assertThat(aliveNodeUrls.size()).isEqualTo(NB_NODES);
+        assertThat(aliveNodeUrls.size()).isEqualTo(TestSSHInfrastructureV2.NB_NODES);
 
         // the recovered nodes should be usable, try to lock/unlock them to see
         BooleanWrapper lockSucceeded = resourceManager.lockNodes(allNodes);
